added more validations to delete and formatted displaying info

+++ Menu.java
@@ -54,7 +54,7 @@ public class Menu {
 
 	public static void addClub() {
 			try{
-				System.out.println("Enter the name of the club");
+				System.out.println("\nEnter the name of the club");
 				String clubName = scanner.nextLine();
 				if(premierLeague.checkForClub(clubName)) {
 					System.out.println("\n" + clubName + " is already in the premier league");
@@ -73,29 +73,38 @@ public class Menu {
 
 	public static void deleteClub() {
 		if(premierLeague.isClubListEmpty()) {
-			System.out.println("No clubs found in the premier league!!! Please add atleast one club to continue");
+			System.out.println("No clubs found in the premier league!!! Please add atleast one club to continue.");
 			return;
 		}
 		premierLeague.displayClubNames();
 		System.out.println("Enter the name of the club to delete from the above list");
 		String clubNameDelete = scanner.nextLine();
 		if(!premierLeague.checkForClub(clubNameDelete)) {
-			System.out.println(clubNameDelete +" does not exist in the premier league");
+			System.out.println(clubNameDelete +" does not exist in the premier league.");
 			return;
 		}
-		premierLeague.deleteClub(clubNameDelete);
+		System.out.println("Warning! You are about to delete " + clubNameDelete + ". Please confirm(y/n).");
+		String userChoice = scanner.nextLine().toLowerCase();
+		switch(userChoice) {
+			case "y":
+				premierLeague.deleteClub(clubNameDelete);
+				System.out.println("Successfully removed " + clubNameDelete + " from the premier league.");
+				break;
+			default:
+				System.out.println("Deletion Aborted! "+ clubNameDelete + " was not removed from the premier league.");
+		}
 	}	
 
 	public static void displayClub() {
 		if(premierLeague.isClubListEmpty()) {
-			System.out.println("No clubs found in the premier league!!! Please add atleast one club to continue");
+			System.out.println("No clubs found in the premier league!!! Please add atleast one club to continue.");
 			return;
 		}
 		premierLeague.displayClubNames();
 		System.out.println("Enter the name of the club from the above list to see statistics");
 		String clubNameFind = scanner.nextLine();
 		if(!premierLeague.checkForClub(clubNameFind)) {
-			System.out.println(clubNameFind +" does not exist in the premier league");
+			System.out.println(clubNameFind +" does not exist in the premier league.");
 			return;
 		}
 		premierLeague.displayStatistics(clubNameFind);
@@ -103,7 +112,7 @@ public class Menu {
 
 	public static void displayTable() {
 		if(premierLeague.isClubListEmpty()) {
-			System.out.println("No clubs found in the premier league!!! Please add atleast one club to continue");
+			System.out.println("No clubs found in the premier league!!! Please add atleast one club to continue.");
 			return;
 		}
 		premierLeague.displayTable();
@@ -118,19 +127,19 @@ public class Menu {
 		System.out.println("\nEnter the home team from the above club list");
 		String teamA = scanner.nextLine();
 		if(!premierLeague.checkForClub(teamA)) {
-			System.out.println(teamA + " does not exist in the premier league");
+			System.out.println(teamA + " does not exist in the premier league.");
 			return;
 		}
 		System.out.println("Enter the away team from the above club list");
 		String teamB = scanner.nextLine();
 		if(!premierLeague.checkForClub(teamB)) {
-			System.out.println(teamB + " does not exist in the premier league");
+			System.out.println(teamB + " does not exist in the premier league.");
 			return;
 		}
 		SportsClub clubA = premierLeague.getClubFromList(teamA);
 		SportsClub clubB = premierLeague.getClubFromList(teamB);
 		if(clubA.equals(clubB)) {
-			System.out.println("Cannot play against the same team!!! Please select two separate teams");
+			System.out.println("Cannot play against the same team!!! Please select two separate teams.");
 			return;
 		}
